Fix bug where item count object wasn't correctly initialized
If the user navigated to the shop page before the items had been
fetched from the external API, the itemValues state would be
initialized to an empty object, since there were no items to set
counts for.

Now, a loading screen is shown if there are no items, and the
item component is only mounted once the items are available.
Therefore, the itemValues object is initialized with zeros
associated with each item ID, and React doesn't complain about an
uncontrolled component becoming controlled.