Add accessibility labels and test suite for ShoppingItems
The labels are necessary to target specific items in the DOM. I
had initially assumed that the screenreader experience would be
better having the input / buttons presented after hearing the item
without the item name repeated every time -- particularly tiresome
if the item name is long. If this was a real app, I'd test it and
look up best practices here. Since it's not, I added the labels and
put the names at the end so users can quickly skip past hearing the
full item name once they've identified the relevant element.

I'm not sure the last test in this suite is really meaningful, but
I couldn't think of / find a better way to change mocked context.